How to play
Created the How to Play scene.
The player can now navigate to and from the How to Play scene with both keyboard/mouse and controller.
Button icons change based on the device being used.
TODO: need to add stuff related to getting points and obstacles.
TODO: need to detect and send the controller type on ready as it loads keyboard first.